refactor: Refatora classe pessoa e coloca atributo idade e função calcularIdade como privados
